Bump jest to latest version, remove objection-transactional-tests
As per https://github.com/smartlyio/objection-transactional-tests/issues/3
