+++ jest.config.js
@@ -8,6 +8,10 @@ module.exports = {
   // The directory where Jest should output its coverage files
   coverageDirectory: 'coverage',
 
+  // A list of paths to modules that run some code to configure or set up
+  // the testing framework before each test file in the suite is executed
+  setupFilesAfterEnv: ['./jest.setup.js'],
+
   // The test environment that will be used for testing
   testEnvironment: 'node'
 }
