+++ package.json
@@ -20,10 +20,9 @@
     "bcrypt": "^5.0.0"
   },
   "devDependencies": {
-    "jest": "^26.4.2",
+    "jest": "^27.0.4",
     "knex": "^0.21.5",
     "objection": "^2.2.3",
-    "objection-transactional-tests": "^0.2.0",
     "sqlite3": "^5.0.0",
     "standard": "^14.3.4"
   },
